+++ lib/src/codecs/base64.dart
@@ -2,62 +2,78 @@
 // All rights reserved. Check LICENSE file for details.
 
 import 'package:hashlib_codecs/src/core/codec.dart';
-import 'package:hashlib_codecs/src/core/converter.dart';
+import 'package:hashlib_codecs/src/core/alphabet_converter.dart';
+
+// ignore: constant_identifier_names
+const int __ = -1;
 
 const _base64Encoding = [
-  65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, //
-  84, 85, 86, 87, 88, 89, 90, 97, 98, 99, 100, 101, 102, 103, 104, 105, 106,
-  107, 108, 109, 110, 111, 112, 113, 114, 115, 116, 117, 118, 119, 120, 121,
-  122, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 43, 47
+  0x41, 0x42, 0x43, 0x44, 0x45, 0x46, 0x47, 0x48, //
+  0x49, 0x4a, 0x4b, 0x4c, 0x4d, 0x4e, 0x4f, 0x50,
+  0x51, 0x52, 0x53, 0x54, 0x55, 0x56, 0x57, 0x58,
+  0x59, 0x5a, 0x61, 0x62, 0x63, 0x64, 0x65, 0x66,
+  0x67, 0x68, 0x69, 0x6a, 0x6b, 0x6c, 0x6d, 0x6e,
+  0x6f, 0x70, 0x71, 0x72, 0x73, 0x74, 0x75, 0x76,
+  0x77, 0x78, 0x79, 0x7a, 0x30, 0x31, 0x32, 0x33,
+  0x34, 0x35, 0x36, 0x37, 0x38, 0x39, 0x2b, 0x2f,
 ];
 
 const _base64urlEncoding = [
-  65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, //
-  84, 85, 86, 87, 88, 89, 90, 97, 98, 99, 100, 101, 102, 103, 104, 105, 106,
-  107, 108, 109, 110, 111, 112, 113, 114, 115, 116, 117, 118, 119, 120, 121,
-  122, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 45, 95
+  0x41, 0x42, 0x43, 0x44, 0x45, 0x46, 0x47, 0x48, //
+  0x49, 0x4a, 0x4b, 0x4c, 0x4d, 0x4e, 0x4f, 0x50,
+  0x51, 0x52, 0x53, 0x54, 0x55, 0x56, 0x57, 0x58,
+  0x59, 0x5a, 0x61, 0x62, 0x63, 0x64, 0x65, 0x66,
+  0x67, 0x68, 0x69, 0x6a, 0x6b, 0x6c, 0x6d, 0x6e,
+  0x6f, 0x70, 0x71, 0x72, 0x73, 0x74, 0x75, 0x76,
+  0x77, 0x78, 0x79, 0x7a, 0x30, 0x31, 0x32, 0x33,
+  0x34, 0x35, 0x36, 0x37, 0x38, 0x39, 0x2d, 0x5f,
 ];
 
 const _base64Decoding = [
-  -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, //
-  -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2, -2,
-  -2, -2, -2, -2, -2, 62, -2, 62, -2, 63, 52, 53, 54, 55, 56, 57, 58, 59, 60,
-  61, -2, -2, -2, -1, -2, -2, -2, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13,
-  14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, -2, -2, -2, -2, 63, -2, 26,
-  27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45,
-  46, 47, 48, 49, 50, 51
+  __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, //
+  __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __,
+  __, __, __, __, __, 62, __, 62, __, 63, 52, 53, 54, 55, 56, 57, 58, 59, 60,
+  61, __, __, __, __, __, __, __, 00, 01, 02, 03, 04, 05, 06, 07, 08, 09, 10,
+  11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, __, __, __, __,
+  63, __, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42,
+  43, 44, 45, 46, 47, 48, 49, 50, 51, __, __, __, __, __, __, __, __, __, __,
 ];
 
-class B64Codec extends Uint8Codec {
+// ========================================================
+// Base-64 Codec
+// ========================================================
+
+class Base64Codec extends ByteCodec {
   @override
-  final Uint8Encoder encoder;
+  final AlphabetEncoder encoder;
 
   @override
-  final decoder = const Uint8Decoder(
+  final decoder = const AlphabetDecoder(
     bits: 6,
+    padding: 61,
     alphabet: _base64Decoding,
   );
 
-  /// Codec instance to encode and decode 8-bit integer sequence to 6-bit Base64
-  /// character sequence.
-  const B64Codec()
-      : encoder = const Uint8Encoder(
+  /// Codec instance to encode and decode 8-bit integer sequence to 6-bit
+  /// Base-64 character sequence.
+  const Base64Codec()
+      : encoder = const AlphabetEncoder(
           bits: 6,
           alphabet: _base64Encoding,
         );
 
   /// Codec instance to encode and decode 8-bit integer sequence to a modified
-  /// 6-bit Base64 character sequence that is both URL and filename safe.
-  const B64Codec.url()
-      : encoder = const Uint8Encoder(
+  /// 6-bit Base-64 character sequence that is both URL and filename safe.
+  const Base64Codec.url()
+      : encoder = const AlphabetEncoder(
           bits: 6,
           alphabet: _base64urlEncoding,
         );
 
   /// Creates a constructor where the encoder will use character `=` as padding,
   /// which is appended at the end out the output to fill up any partial bytes.
-  const B64Codec.padded()
-      : encoder = const Uint8Encoder(
+  const Base64Codec.padded()
+      : encoder = const AlphabetEncoder(
           bits: 6,
           padding: 61,
           alphabet: _base64Encoding,
@@ -65,8 +81,8 @@ class B64Codec extends Uint8Codec {
 
   /// Codec instance where the encoder will use character `=` as padding,
   /// which is appended at the end out the output to fill up any partial bytes.
-  const B64Codec.urlpadded()
-      : encoder = const Uint8Encoder(
+  const Base64Codec.urlpadded()
+      : encoder = const AlphabetEncoder(
           bits: 6,
           padding: 61,
           alphabet: _base64urlEncoding,