+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# 2.0.0
+
+- Generalize encoding and decoding
+- Adds Base-8 (Octal) codec support
+  - New class: `Base8Codec`
+  - New methods: `fromOctal`, `toOctal`
+  - New constant: `base8`
+- Renames a lot of exports
+  - `Uint8Converter` -> `BitConverter`
+  - `Uint8Codec` -> `ByteCodec`
+  - `BinaryCodec` -> `Base2Codec`
+  - `B16Codec` -> `Base16Codec`
+  - `B64Codec` -> `Base64Codec`
+- Change internal methods of `ByteCodec`
+- Separates base encoder and decoders
+  - Generic encoders: `BitEncoder`, `AlphabetEncoder`
+  - Generic decoders: `BitDecoder`, `AlphabetDecoder`
+- Exports two new types:
+  - `BigIntEncoder`
+  - `BigIntDecoder`
+
 # 1.2.0
 
 - Adds BigInt codec support.